move makeLinearCorrelatedLabel out of src and into TestUtils
