+++ src/main/scala/io/citrine/lolo/stats/StatsUtils.scala
@@ -1,9 +1,5 @@
 package io.citrine.lolo.stats
 
-import io.citrine.lolo.linear.LinearRegressionLearner
-
-import scala.util.Random
-
 object StatsUtils {
 
   /** Compute the mean of a (weighted) vector, X */
@@ -42,29 +38,4 @@ object StatsUtils {
     if (sigma2X == 0 || sigma2Y == 0) return 0.0
     covariance(X, Y, Some(actualWeights)) / math.sqrt(sigma2X * sigma2Y)
   }
-
-  /** Given a univariate data set, construct a corresponding data set with the desired Pearson correlation coefficient.
-    * The procedure is to generate new data randomly, calculate the residuals of a linear regression,
-    * and then create a suitable linear combination of X and the residuals.
-    * Source: https://stats.stackexchange.com/a/313138
-    *
-    * @param X    sequence of values
-    * @param rho  desired Pearson correlation coefficient
-    * @param rng  random number generator
-    * @return     sequence of values that have desired correlation with X
-    */
-  def makeLinearCorrelatedData(X: Seq[Double], rho: Double, rng: Random = new Random()): Seq[Double] = {
-    require(rho >= -1.0 && rho <= 1.0, "correlation coefficient must be between -1.0 and 1.0")
-    val Y = Seq.fill(X.length)(rng.nextGaussian())
-    val linearLearner = LinearRegressionLearner()
-    val linearModel = linearLearner.train(X.zip(Y).map { case (x, y) => (Vector(x), y) } ).getModel()
-    val yPred = linearModel.transform(X.map(Vector(_))).getExpected()
-    val residuals = Y.zip(yPred).map { case (actual, predicted) => actual - predicted }
-    val muX = mean(X)
-    val stdX = math.sqrt(variance(X))
-    val muResiduals = mean(residuals)
-    val stdResiduals = math.sqrt(variance(residuals))
-    X.zip(residuals).map { case (x, residual) => rho * stdResiduals * x + math.sqrt(1 - rho * rho) * stdX * residual}
-  }
-
 }
