+++ src/test/scala/io/citrine/lolo/TestUtils.scala
@@ -1,11 +1,11 @@
 package io.citrine.lolo
 
 import breeze.stats.distributions.{RandBasis, ThreadLocalRandomGenerator}
+import io.citrine.lolo.linear.LinearRegressionLearner
+import io.citrine.lolo.stats.StatsUtils.variance
 import io.citrine.lolo.stats.functions.Friedman
 import org.apache.commons.math3.random.MersenneTwister
 
-import scala.collection.generic.CanBuildFrom
-import scala.collection.mutable
 import scala.util.{Random, Try}
 
 /**
@@ -44,6 +44,28 @@ object TestUtils {
       val input = Vector.fill(cols)(xscale * rnd.nextDouble() + xoff)
       (input, function(input) + noise * rnd.nextGaussian())
     }
+  }
+
+  /** Given a univariate data set, construct a corresponding data set with the desired Pearson correlation coefficient.
+    * The procedure is to generate new data randomly, calculate the residuals of a linear regression,
+    * and then create a suitable linear combination of X and the residuals.
+    * Source: https://stats.stackexchange.com/a/313138
+    *
+    * @param X    sequence of values
+    * @param rho  desired Pearson correlation coefficient
+    * @param rng  random number generator
+    * @return     sequence of values that have desired correlation with X
+    */
+  def makeLinearCorrelatedData(X: Seq[Double], rho: Double, rng: Random = new Random()): Seq[Double] = {
+    require(rho >= -1.0 && rho <= 1.0, "correlation coefficient must be between -1.0 and 1.0")
+    val Y = Seq.fill(X.length)(rng.nextGaussian())
+    val linearLearner = LinearRegressionLearner()
+    val linearModel = linearLearner.train(X.zip(Y).map { case (x, y) => (Vector(x), y) } ).getModel()
+    val yPred = linearModel.transform(X.map(Vector(_))).getExpected()
+    val residuals = Y.zip(yPred).map { case (actual, predicted) => actual - predicted }
+    val stdX = math.sqrt(variance(X))
+    val stdResiduals = math.sqrt(variance(residuals))
+    X.zip(residuals).map { case (x, residual) => rho * stdResiduals * x + math.sqrt(1 - rho * rho) * stdX * residual}
   }
 
   def iterateTrainingData(
